:recycle: Refactor: User 도메인 이동
<body>
- user/domain : entity class
- user/repository : repository
- user/type : enum 등의 type class

<footer>
- 관련: #139
