+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/application/MoimMemoFacade.java
@@ -25,7 +25,7 @@ import com.namo.spring.application.external.domain.group.ui.dto.GroupDiaryReques
 import com.namo.spring.application.external.domain.group.ui.dto.GroupDiaryResponse;
 import com.namo.spring.application.external.domain.group.ui.dto.GroupScheduleRequest;
 import com.namo.spring.application.external.domain.user.application.impl.UserService;
-import com.namo.spring.application.external.domain.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.domain.User;
 import com.namo.spring.application.external.global.common.constant.FilePath;
 import com.namo.spring.application.external.global.utils.FileUtils;
