+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/application/converter/MoimScheduleResponseConverter.java
@@ -11,7 +11,7 @@ import com.namo.spring.application.external.domain.group.domain.MoimScheduleAndU
 
 import com.namo.spring.application.external.domain.individual.domain.Schedule;
 
-import com.namo.spring.application.external.domain.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.domain.User;
 
 public class MoimScheduleResponseConverter {
 	private MoimScheduleResponseConverter() {
