+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/application/impl/MoimMemoLocationService.java
@@ -12,7 +12,7 @@ import com.namo.spring.application.external.domain.group.domain.MoimSchedule;
 import com.namo.spring.application.external.domain.group.repository.diary.MoimMemoLocationAndUserRepository;
 import com.namo.spring.application.external.domain.group.repository.diary.MoimMemoLocationImgRepository;
 import com.namo.spring.application.external.domain.group.repository.diary.MoimMemoLocationRepository;
-import com.namo.spring.application.external.domain.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.domain.User;
 import com.namo.spring.core.common.code.status.ErrorStatus;
 import com.namo.spring.core.common.exception.GroupException;
