+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/domain/Moim.java
@@ -17,7 +17,7 @@ import jakarta.persistence.Table;
 
 import com.namo.spring.application.external.domain.group.domain.constant.MoimStatus;
 
-import com.namo.spring.application.external.domain.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.domain.User;
 
 import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
