+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/domain/MoimMemoLocationAndUser.java
@@ -10,7 +10,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import jakarta.persistence.Table;
 
-import com.namo.spring.application.external.domain.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.domain.User;
 
 import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
