+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/domain/MoimScheduleAndUser.java
@@ -20,7 +20,7 @@ import com.namo.spring.application.external.domain.group.domain.constant.Visible
 
 import com.namo.spring.application.external.domain.individual.domain.Category;
 
-import com.namo.spring.application.external.domain.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.domain.User;
 
 import com.namo.spring.db.mysql.common.model.BaseTimeEntity;
