+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/group/repository/group/MoimAndUserRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.Query;
 import com.namo.spring.application.external.domain.group.domain.Moim;
 import com.namo.spring.application.external.domain.group.domain.MoimAndUser;
 
-import com.namo.spring.application.external.domain.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.domain.User;
 
 public interface MoimAndUserRepository extends JpaRepository<MoimAndUser, Long> {
