+++ application/external-api/src/main/java/com/namo/spring/application/external/domain/individual/application/ScheduleFacade.java
@@ -28,7 +28,7 @@ import com.namo.spring.application.external.domain.individual.domain.constant.Pe
 import com.namo.spring.application.external.domain.individual.ui.dto.ScheduleRequest;
 import com.namo.spring.application.external.domain.individual.ui.dto.ScheduleResponse;
 import com.namo.spring.application.external.domain.user.application.impl.UserService;
-import com.namo.spring.application.external.domain.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.domain.User;
 import com.namo.spring.application.external.global.common.constant.FilePath;
 import com.namo.spring.application.external.global.utils.FileUtils;
