--- application/external-api/src/main/java/com/namo/spring/application/external/domain/user/domain/constant/Content.java
@@ -1,5 +1,0 @@
-package com.namo.spring.application.external.domain.user.domain.constant;
-
-public enum Content {
-	TERMS_OF_USE, PERSONAL_INFORMATION_COLLECTION
-}
