--- application/external-api/src/main/java/com/namo/spring/application/external/domain/user/domain/constant/SocialType.java
@@ -1,5 +1,0 @@
-package com.namo.spring.application.external.domain.user.domain.constant;
-
-public enum SocialType {
-	KAKAO, NAVER, APPLE
-}
