--- application/external-api/src/main/java/com/namo/spring/application/external/domain/user/domain/constant/UserStatus.java
@@ -1,5 +1,0 @@
-package com.namo.spring.application.external.domain.user.domain.constant;
-
-public enum UserStatus {
-	ACTIVE, INACTIVE
-}
