+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/user/repository/TermRepository.java
@@ -1,13 +1,13 @@
-package com.namo.spring.application.external.domain.user.repository;
+package com.namo.spring.db.mysql.domains.user.repository;
 
 import java.util.List;
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import com.namo.spring.application.external.domain.user.domain.Term;
-import com.namo.spring.application.external.domain.user.domain.User;
-import com.namo.spring.application.external.domain.user.domain.constant.Content;
+import com.namo.spring.db.mysql.domains.user.domain.Term;
+import com.namo.spring.db.mysql.domains.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.type.Content;
 
 public interface TermRepository extends JpaRepository<Term, Long> {
 	Optional<Term> findTermByContentAndUser(Content content, User user);
