+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/user/repository/UserRepository.java
@@ -1,4 +1,4 @@
-package com.namo.spring.application.external.domain.user.repository;
+package com.namo.spring.db.mysql.domains.user.repository;
 
 import java.time.LocalDateTime;
 import java.util.List;
@@ -7,9 +7,9 @@ import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 
-import com.namo.spring.application.external.domain.user.domain.User;
-import com.namo.spring.application.external.domain.user.domain.constant.SocialType;
-import com.namo.spring.application.external.domain.user.domain.constant.UserStatus;
+import com.namo.spring.db.mysql.domains.user.domain.User;
+import com.namo.spring.db.mysql.domains.user.type.SocialType;
+import com.namo.spring.db.mysql.domains.user.type.UserStatus;
 
 public interface UserRepository extends JpaRepository<User, Long> {
 	Optional<User> findUserByEmail(String email);
