+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/user/type/Content.java
@@ -1,0 +1,5 @@
+package com.namo.spring.db.mysql.domains.user.type;
+
+public enum Content {
+	TERMS_OF_USE, PERSONAL_INFORMATION_COLLECTION
+}
