+++ storage/db-mysql/src/main/java/com/namo/spring/db/mysql/domains/user/type/UserStatus.java
@@ -1,0 +1,5 @@
+package com.namo.spring.db.mysql.domains.user.type;
+
+public enum UserStatus {
+	ACTIVE, INACTIVE
+}
